Create local mode worker stub and split local and remote handlers
